Organize socket connection
