+++ bedwars-plugin/src/main/java/com/tomkeuper/bedwars/lobbyconnection/LoadedUser.java
@@ -18,7 +18,7 @@
  * Contact e-mail: andrew.dascalu@gmail.com
  */
 
-package com.tomkeuper.bedwars.lobbysocket;
+package com.tomkeuper.bedwars.lobbyconnection;
 
 import com.tomkeuper.bedwars.BedWars;
 import com.tomkeuper.bedwars.api.configuration.ConfigPath;
